Update moved procedures to have new class names
