+++ extended/src/main/java/apoc/load/LoadJsonExtended.java
@@ -46,8 +46,7 @@ public class LoadJsonExtended {
     public URLAccessChecker urlAccessChecker;
 
     @SuppressWarnings("unchecked")
-    @Procedure(name = "apoc.load.jsonParams", deprecatedBy = "This procedure is being moved to APOC Extended.")
-    @Deprecated
+    @Procedure(name = "apoc.load.jsonParams")
     @QueryLanguageScope(scope = {QueryLanguage.CYPHER_25})
     @Description(
             "Loads parameters from a JSON URL (e.g. web-API) as a stream of values if the given JSON file is a `LIST<ANY>`.\n"
